+++ pkg/mousemover/mouseMover.go
@@ -6,9 +6,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/sirupsen/logrus"
-	log "github.com/sirupsen/logrus"
-
 	"github.com/go-vgo/robotgo"
 	"github.com/prashantgupta24/activity-tracker/pkg/activity"
 	"github.com/prashantgupta24/activity-tracker/pkg/tracker"
@@ -27,7 +24,7 @@ type MouseMover struct {
 const (
 	timeout     = 100 //ms
 	logDir      = "log"
-	logFileName = "logFile-amm-1"
+	logFileName = "logFile-amm-2"
 )
 
 //Start the main app
@@ -47,11 +44,14 @@ func (m *MouseMover) Start() {
 	}
 
 	heartbeatCh := activityTracker.Start()
-	run(heartbeatCh, m)
+	m.run(heartbeatCh, activityTracker)
 }
 
-func run(heartbeatCh chan *tracker.Heartbeat, m *MouseMover) {
-	go func(m *MouseMover) {
+func (m *MouseMover) run(heartbeatCh chan *tracker.Heartbeat, activityTracker *tracker.Instance) {
+	go func() {
+		if m.isRunning() {
+			return
+		}
 		logger := getLogger(m, false) //set writeToFile=true only for debugging
 		m.updateRunningStatus(true)
 		movePixel := 10
@@ -91,7 +91,7 @@ func run(heartbeatCh chan *tracker.Heartbeat, m *MouseMover) {
 						logger.Errorf("timeout happened after %vms while trying to move mouse", timeout)
 					}
 				} else {
-					logger.Infof("activity detected in the last %v seconds.", int(heartbeatInterval))
+					logger.Infof("activity detected in the last %v seconds.", int(activityTracker.HeartbeatInterval))
 					logger.Infof("Activity type:\n")
 					for activityType, times := range heartbeat.ActivityMap {
 						logger.Infof("activityType : %v times: %v\n", activityType, len(times))
@@ -110,34 +110,7 @@ func run(heartbeatCh chan *tracker.Heartbeat, m *MouseMover) {
 				return
 			}
 		}
-	}(m)
-}
-
-func (m *MouseMover) isRunning() bool {
-	m.mutex.RLock()
-	defer m.mutex.RUnlock()
-	return m.runningStatus
-}
-func (m *MouseMover) updateRunningStatus(isRunning bool) {
-	m.mutex.Lock()
-	defer m.mutex.Unlock()
-	m.runningStatus = isRunning
-}
-
-func moveAndCheck(movePixel int, mouseMoveSuccessCh chan bool) {
-	currentX, currentY := robotgo.GetMousePos()
-	moveToX := currentX + movePixel
-	moveToY := currentY + movePixel
-	robotgo.MoveMouse(moveToX, moveToY)
-
-	//check if mouse moved. Sometimes mac users need to give
-	//extra permission for controlling the mouse
-	movedX, movedY := robotgo.GetMousePos()
-	if movedX == currentX && movedY == currentY {
-		mouseMoveSuccessCh <- false
-	} else {
-		mouseMoveSuccessCh <- true
-	}
+	}()
 }
 
 //Quit the app
@@ -157,32 +130,4 @@ func GetInstance() *MouseMover {
 		instance = &MouseMover{}
 	}
 	return instance
-}
-
-func getLogger(m *MouseMover, doWriteToFile bool) *log.Logger {
-	logger := log.New()
-	logger.Formatter = &logrus.TextFormatter{
-		FullTimestamp: true,
-	}
-
-	if doWriteToFile {
-		_, err := os.Stat(logDir)
-		if err != nil {
-			if os.IsNotExist(err) {
-				err = os.Mkdir(logDir, os.ModePerm)
-				if err != nil {
-					log.Fatalf("error creating dir: %v", err)
-				}
-			}
-		}
-
-		logFile, err := os.OpenFile(logDir+"/"+logFileName, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
-		if err != nil {
-			log.Fatalf("error opening file: %v", err)
-		}
-		logger.SetOutput(logFile)
-		m.logFile = logFile
-	}
-
-	return logger
 }
